Fix mismatching tools and slimboot issue
Universal Payload support in SBL is keeping moving to align with
the latest spec. However, the LinuxPayload enabling was done basing
on the initial spec. So in the instruciton, specific commit needs
to be used in order to reproduce the result.

Also the latest GCC toolchain enabled some new instruction that is
not supported by QEMU. So adjusted the Linux kernel build flags to
ensure the image is bootable on QEMU.

Signed-off-by: Maurice Ma <maurice.ma@intel.com>